pyproject.toml
[tool.poetry]
name = "last-war-butler-bot"
version = "0.1.0"
description = ""
authors = ["georgelee144 <george@ml.cafe>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.6.0"
toml = "^0.10.2"
pendulum = "^3.0.0"
apscheduler = "^3.10.4"
groq = "^0.11.0"
google-generativeai = "^0.8.2"
google-cloud-vision = "^3.7.4"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
